fixing module organization for data parsing methods
